feat(security): 添加 CSP 并补齐 frame-ancestors 响应头
- web/index.html: 添加 CSP meta，收紧资源加载策略；主题首帧设置迁移到独立 module 启动脚本，避免依赖 unsafe-inline。
- electron: 通过 webRequest.onHeadersReceived 注入响应头，补齐 `frame-ancestors 'none'` 与 `X-Frame-Options: DENY`。
- web: 抽取 App.tsx 的公用常量/组件/工具到 app-shared.tsx，降低入口文件复杂度。
- test: `npm test` 先跑 web typecheck（新增 web/tsconfig.typecheck.json）。
- fix: history-find 移除 createTreeWalker 过时参数，避免 any。
- deps: package-lock 中 caniuse-lite 小版本更新。
